[Backend] delete all related data on removing user, host, guest and listing row

+++ backend/src/controllers/ListingController.ts
@@ -121,21 +121,11 @@ export const deleteListing = async (req: Request, res: Response) => {
 		const { listingId } = req.params;
 
 		const listing = await prisma.$transaction([
-			prisma.listingPhoto.deleteMany({
-				where: {
-					listingId,
-				},
-			}),
-			prisma.address.deleteMany({
-				where: {
-					ownerId: listingId,
-				},
-			}),
-			prisma.listing.delete({
-				where: {
-					listingId,
-				},
-			}),
+			prisma.address.deleteMany({ where: { ownerId: listingId } }),
+			prisma.listingPhoto.deleteMany({ where: { listingId } }),
+			prisma.reservation.deleteMany({ where: { listingId } }),
+
+			prisma.listing.delete({ where: { listingId } }),
 		]);
 
 		res.status(200).json(listing);
